Python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "whisper>=1.1.10",
]
